fix WithContext race when parent already canceled
if WithContext is called with a canceled parent context, there is a race
between the internal call to addChild() and the cancelation checking
goroutine calling Kill().

This change moves the launching of the goroutine until after addChild()
has been called. Alternatively, the call to addChild could done under
the tomb mutex.